First attempt to add uint64_t texture for ids

+++ src/Engine/Assets/Shaders/Renderer2D_CircleShader.frag
@@ -1,13 +1,15 @@
-#version 450 core
+#version 460
+
+#extension GL_ARB_gpu_shader_int64 : require
 
 layout(location = 0) in vec4 fragColor;
 layout(location = 1) in vec2 fragTexCoord;
 layout(location = 2) in float fragThickness;
 layout(location = 3) in float fragFade;
-layout(location = 4) in flat int fragEntityID;
+layout(location = 4) in flat uint64_t fragEntityID;
 
 layout(location = 0) out vec4 outColor;
-layout(location = 1) out float outEntityID;
+layout(location = 1) out uint64_t outEntityID;
 
 void main()
 {
